+++ packages/core/CHANGELOG.md
@@ -4,7 +4,7 @@ name: Changelog
 description: All notable changes to this project will be documented in this file.
 ---
 
-## [Unreleased]
+## [0.2.0] - 2024-03-31
 
 ## Added